refactor(acp): extract runtime composition
