+++ src/main/acp/ipc.ts
@@ -1,5 +1,4 @@
 import { createHmac, randomBytes, randomUUID } from 'node:crypto'
-import { homedir } from 'node:os'
 
 import { app } from 'electron'
 
@@ -14,54 +13,27 @@ import type {
   AcpCompactSessionRequest,
   AcpConnectRequest,
   AcpCreateSessionRequest,
-  AcpRuntimeEvent,
   AcpDeleteSessionRequest,
-  AcpPermissionRequest,
   AcpPermissionResponse,
   AcpPromptRequest,
   AcpResumeSessionRequest,
   AcpRevokePermissionGrantRequest,
-  AcpSetPermissionProfileRequest,
-  AcpStateSnapshot
+  AcpSetPermissionProfileRequest
 } from '../../shared/acp'
 import { DEFAULT_ARTIFACT_PROJECT_NAME } from '../../shared/artifacts'
-import { AcpRuntime, type AcpRuntimeCallbacks } from './runtime'
 import { AcpRuntimeCoordinator } from './runtime-coordinator'
-import {
-  createAcpCreateSessionWorkflow,
-  type AcpCreateSessionWorkflow
-} from './create-session-workflow'
+import type { AcpCreateSessionWorkflow } from './create-session-workflow'
 import { installAgentShutdownGuard } from './shutdown-guard'
-import { AgentMcpHttpHost } from './mcp-http-host'
 import { ArtifactRepository } from '../artifacts/repository'
-import type { ArtifactProvenanceRepository } from '../artifacts/provenance-repository'
-import type { ArtifactRunRegistry } from '../artifacts/run-registry'
-import {
-  runTaskNotificationInBackground,
-  type TaskNotificationService
-} from '../notifications/task-notifications'
-import { NotebookLocalRpcServer } from '../notebook/local-rpc-server'
+import type { TaskNotificationService } from '../notifications/task-notifications'
 import { NotebookRunRepository } from '../notebook/repository'
 import {
   NotebookRuntimeService,
   type NotebookRuntimeServiceOptions
 } from '../notebook/runtime-service'
 import { resolveConfigRoot, resolveDataRoot } from '../storage-root'
-import type { AcpSettingsCapabilities } from '../settings/service-capabilities'
-import type { UploadRepository } from '../uploads/repository'
-import type { PermissionGrantRegistry } from '../permission-grants/registry'
-import { projectRegistrySessionGrants } from './permission-broker'
 import { broadcastToRenderers } from '../renderer-broadcast'
 import { createLogger, diagnosticErrorFields, errorLogFields } from '../logger'
-import type { ProfileService } from '../specialist/service'
-import {
-  buildSpecialistIdentityAppend,
-  buildSpecialistIdentityPrefix
-} from '../specialist/identity'
-import {
-  resolveEffectiveSpecialistSkills,
-  filterSpecialistConnectorSkills
-} from '../../shared/specialist'
 
 const log = createLogger('acp')
 const resumeLogHashKey = randomBytes(32)
@@ -130,236 +102,8 @@ const logResumeDiagnostic = (
   }
 }
 
-type AcpIpcArtifacts = {
-  repository: ArtifactRepository
-  runRegistry: ArtifactRunRegistry
-  provenanceRepository?: Pick<
-    ArtifactProvenanceRepository,
-    'listRunVersions' | 'writeAppGeneratedVersion'
-  > &
-    Partial<Pick<ArtifactProvenanceRepository, 'resolveVersionContent'>>
-}
-
-type AcpIpcOptions = AcpIpcArtifacts & {
-  mcpEntryPath: string
-  uploadRepository: UploadRepository
-  notebookRpcServer: NotebookLocalRpcServer
-  authorizeSkillImportReferencedUploads: (
-    projectId: string,
-    sessionId: string,
-    paths: string[]
-  ) => Promise<() => void>
-  // Drives the agent spawn env from the active provider so switching takes effect on reconnect.
-  settingsService: AcpSettingsCapabilities
-  permissionGrantRegistry?: PermissionGrantRegistry
-  initializationBarrier?: Promise<unknown>
-  // Observes prompt starts and terminal turn events for desktop notifications. Optional so tests
-  // and headless setups can run without a notification surface.
-  taskNotifications?: TaskNotificationService
-  onSessionTurnStarted?: (sessionId: string, turnToken: string) => void
-  onSessionTurnEnded?: (sessionId: string, turnToken: string) => void
-  onSkillImportAttachmentEligible?: (
-    sessionId: string,
-    turnToken: string,
-    attachmentUri: string
-  ) => void
-  onSessionCancellationRequested?: (sessionId: string) => void
-  onSessionUnavailable?: (sessionId: string) => void
-  onAllSessionsCancellationRequested?: () => void
-  onDisconnected?: () => void
-  beforeSessionDelete?: (sessionId: string) => Promise<void>
-  // Provides fresh Specialist Profiles for first-turn identity injection. Optional so existing
-  // tests and headless setups that construct the runtime without specialist support are unaffected.
-  profileService?: ProfileService
-}
-
 // Sends one runtime payload through the typed application-event compatibility facade.
 const broadcast = broadcastToRenderers
-
-// Creates the runtime coordinator used by all ACP IPC handlers and artifact claims. Each child runtime
-// captures one framework generation so sessions on different frameworks can remain live concurrently.
-const createAcpRuntime = ({
-  mcpEntryPath,
-  repository,
-  runRegistry,
-  provenanceRepository,
-  uploadRepository,
-  notebookRpcServer,
-  authorizeSkillImportReferencedUploads,
-  settingsService,
-  permissionGrantRegistry,
-  initializationBarrier,
-  taskNotifications,
-  onSessionTurnStarted,
-  onSessionTurnEnded,
-  onSkillImportAttachmentEligible,
-  onSessionCancellationRequested,
-  onSessionUnavailable,
-  onAllSessionsCancellationRequested,
-  onDisconnected,
-  beforeSessionDelete,
-  profileService
-}: AcpIpcOptions): AcpRuntimeCoordinator => {
-  const configRoot = resolveConfigRoot()
-  const dataRoot = resolveDataRoot()
-  const defaultCwd = homedir()
-  const callbacks: AcpRuntimeCallbacks = {
-    onStateChanged: (state: AcpStateSnapshot) => broadcast('acp:state', state),
-    onEvent: (event: AcpRuntimeEvent) => {
-      broadcast('acp:event', event)
-      // Fire-and-forget: a notification hiccup must never stall the renderer event stream.
-      if (taskNotifications) {
-        runTaskNotificationInBackground(
-          () => taskNotifications.handleRuntimeEvent(event),
-          (error) => log.warn('task notification event failed', errorLogFields(error))
-        )
-      }
-    },
-    onPermissionRequest: (request: AcpPermissionRequest) => {
-      broadcast('acp:permission-request', request)
-      // A pending approval parks the turn; an unfocused user gets a desktop nudge.
-      if (taskNotifications) {
-        runTaskNotificationInBackground(
-          () => taskNotifications.handlePermissionRequest(request),
-          (error) => log.warn('permission notification failed', errorLogFields(error))
-        )
-      }
-    }
-  }
-
-  return new AcpRuntimeCoordinator(
-    (runtimeCallbacks, permissionGrantStore) => {
-      // Capture only the non-secret selection per generation. Credentials are resolved fresh at spawn
-      // and released by AcpRuntime after authentication instead of living in this coordinator closure.
-      const selection = settingsService.captureActiveAgentBackendSelection()
-      return new AcpRuntime({
-        appVersion: app.getVersion(),
-        // Packaged macOS apps often start with cwd at "/" or the app bundle; use home instead.
-        defaultCwd,
-        resolveBackend: async (context) =>
-          settingsService.resolveAgentBackend(await selection, context),
-        // HTTP MCP registrations are runtime-owned. Sharing one host would let an old runtime teardown
-        // clear routes belonging to sessions on the newly selected framework.
-        mcpHttpHost: new AgentMcpHttpHost(),
-        skills: {
-          needForceLoad: (ids) => settingsService.skillsNeedingForceLoad(ids),
-          namesForIds: (ids) => settingsService.skillNudgeNamesForIds(ids),
-          descriptorsForIds: (ids, codexHome) =>
-            settingsService.codexSkillDescriptorsForIds(ids, codexHome),
-          catalogForCodexHome: (codexHome) => settingsService.codexSkillCatalog(codexHome)
-        },
-        artifacts: {
-          configRoot,
-          dataRoot,
-          projectName: DEFAULT_ARTIFACT_PROJECT_NAME,
-          mcpEntryPath,
-          repository,
-          runRegistry,
-          provenance: provenanceRepository,
-          getRpcConnection: () => notebookRpcServer.ensureStarted(),
-          issueRpcCapability: (binding) => notebookRpcServer.issueArtifactRunCapability(binding),
-          revokeRpcCapability: (token) => notebookRpcServer.revokeArtifactRunCapability(token)
-        },
-        uploads: { repository: uploadRepository },
-        notebook: {
-          projectName: DEFAULT_ARTIFACT_PROJECT_NAME,
-          mcpEntryPath,
-          getRpcConnection: ({ sessionId, projectId }) =>
-            notebookRpcServer.issueSessionConnection(sessionId, projectId),
-          registerSessionAlias: (aliasSessionId, sessionId) =>
-            notebookRpcServer.registerSessionAlias(aliasSessionId, sessionId),
-          releaseSessionCapabilities: (sessionId) =>
-            notebookRpcServer.releaseSessionCapabilities(sessionId),
-          registerSessionSpecialist: (sessionId, specialistId) =>
-            notebookRpcServer.registerSessionSpecialist(sessionId, specialistId),
-          setArtifactProvenanceContext: (sessionId, context) =>
-            notebookRpcServer.setArtifactProvenanceContext(sessionId, context),
-          registerTurnInputs: (request) => notebookRpcServer.registerNotebookTurnInputs(request)
-        },
-        skillImport: {
-          mcpEntryPath,
-          isEnabled: () => settingsService.getConversationSkillImportEnabled(),
-          getRpcConnection: () => notebookRpcServer.ensureStarted(),
-          registerSessionAlias: (aliasSessionId, sessionId) =>
-            notebookRpcServer.registerSessionAlias(aliasSessionId, sessionId),
-          authorizeReferencedUploads: authorizeSkillImportReferencedUploads
-        },
-        callbacks: runtimeCallbacks,
-        permissionGrantStore,
-        permissionGrantRegistry,
-        // Main process resolves the latest Profile so the renderer never needs to send systemPrompt.
-        resolveSpecialistIdentity: profileService
-          ? async (specialistId: string, frameworkId: string) => {
-              let profile
-              try {
-                profile = await profileService.getById(specialistId)
-              } catch {
-                // Profile not found or corrupt
-                return undefined
-              }
-              if (!profile.enabled) return undefined
-              const append = buildSpecialistIdentityAppend(profile)
-              const prefix = buildSpecialistIdentityPrefix(profile)
-              // For Claude, only the append matters (session-level meta).
-              // For Codex/OpenCode, only the prefix matters (per-turn).
-              // Return both so the runtime can choose by framework.
-              if (frameworkId === 'claude-code') return { append, prefix: '' }
-              return { append: '', prefix }
-            }
-          : undefined,
-        resolveSpecialistSkills: profileService
-          ? async (specialistId) => {
-              try {
-                const profile = await profileService.getById(specialistId)
-                if (!profile.enabled) {
-                  return { kind: 'unavailable', reason: 'The bound specialist is disabled.' }
-                }
-                const effective = resolveEffectiveSpecialistSkills(
-                  profile,
-                  await settingsService.listSpecialistSkillCatalog()
-                )
-                if (effective.kind === 'specialist') {
-                  // Connector tools are delivered as mcp-<id> skills on disk. Claude's options.skills
-                  // whitelist is restrictive, so without these names the agent cannot load the connector
-                  // skill docs and never discovers the tools. Include only the connectors this specialist
-                  // is allowed to use (selectedCapabilities.connectorIds, or all minus full-access
-                  // exclusions) so discovery matches the capability scope; the per-call ConnectorService
-                  // gate still enforces the same config at call time.
-                  const provisioned = await settingsService.provisionedConnectorSkillNames()
-                  const connectorSkills = filterSpecialistConnectorSkills(provisioned, profile)
-                  if (connectorSkills.length > 0) {
-                    return {
-                      ...effective,
-                      frameworkNames: [...effective.frameworkNames, ...connectorSkills]
-                    }
-                  }
-                }
-                return effective
-              } catch {
-                return { kind: 'unavailable', reason: 'The bound specialist is unavailable.' }
-              }
-            }
-          : undefined
-      })
-    },
-    callbacks,
-    defaultCwd,
-    initializationBarrier,
-    onDisconnected,
-    onSessionUnavailable,
-    {
-      onSessionTurnStarted,
-      onSessionTurnEnded,
-      onSkillImportAttachmentEligible,
-      onSessionCancellationRequested,
-      onAllSessionsCancellationRequested,
-      beforeSessionDelete
-    },
-    permissionGrantRegistry
-      ? () => projectRegistrySessionGrants(permissionGrantRegistry.listCached())
-      : undefined
-  )
-}
 
 const registerAcpIpcHandlerSet = (
   runtime: AcpRuntimeCoordinator,
@@ -471,13 +215,6 @@ const installAcpIpcHandlers = (
   }
 }
 
-// Compatibility wrapper for isolated callers; application composition uses the two-phase seam above.
-const registerAcpIpcHandlers = (options: AcpIpcOptions): AcpRuntimeCoordinator => {
-  const runtime = createAcpRuntime(options)
-  installAcpIpcHandlers(runtime, createAcpCreateSessionWorkflow(runtime), options.taskNotifications)
-  return runtime
-}
-
 // Creates the shared notebook runtime used by both renderer IPC and agent MCP calls.
 type DefaultNotebookRuntimeServiceDeps = Pick<
   NotebookRuntimeServiceOptions,
@@ -512,10 +249,4 @@ const createDefaultNotebookRuntimeService = (
   })
 }
 
-export {
-  createAcpRuntime,
-  createDefaultNotebookRuntimeService,
-  installAcpIpcHandlers,
-  registerAcpIpcHandlers
-}
-export type { AcpIpcOptions }
+export { createDefaultNotebookRuntimeService, installAcpIpcHandlers }
